Allow the Help auto-response filter to be used in all rooms
