+++ CODEOWNERS
@@ -4,13 +4,13 @@ data/mods/ssb/ @HoeenCoder
 data/random-teams.ts @TheImmortal
 data/text/ @Marty-D
 databases/ @monsanto
-server/chat-plugins/help.ts @mia-pi-git
 server/chat-plugins/hosts.ts @AnnikaCodes
 server/chat-plugins/mafia.ts @HoeenCoder
 server/chat-plugins/othermetas.ts @KrisXV @TheImmortal
 server/chat-plugins/quotes.ts @mia-pi-git @KrisXV
 server/chat-plugins/random-battles.ts @KrisXV @TheImmortal
 server/chat-plugins/repeats.ts @AnnikaCodes
+server/chat-plugins/responder.ts @mia-pi-git
 server/chat-plugins/rock-paper-scissors.ts @mia-pi-git
 server/chat-plugins/scavenger*.ts @xfix @sparkychildcharlie
 server/chat-plugins/the-studio.ts @KrisXV
